refactor!: rename unplugin exports to feature-oriented names
BREAKING CHANGE: unplugin subpaths and the symbols they export have been
renamed to describe what each plugin does.

Strip transform (build-time PURE/NODE_ENV wrap):

```diff
- import { nostics } from 'nostics/lib'
+ import { nosticsStrip } from 'nostics/unplugin/strip-transform'

  export default defineConfig({
-   plugins: [nostics.vite()],
+   plugins: [nosticsStrip.vite()],
  })
```

Dev-server collector (Vite-only WS forwarder to log file):

```diff
- import { nosticsServer } from 'nostics/dev'
+ import { nosticsCollector } from 'nostics/unplugin/dev-server-collector'

  export default defineConfig({
-   plugins: [nosticsServer.vite()],
+   plugins: [nosticsCollector.vite()],
  })
```

Type renames:

- \`NosticsPluginOptions\` -> \`NosticsStripOptions\`
- \`NosticsServerOptions\` -> \`NosticsCollectorOptions\`

\`nosticsCollector\` is now explicitly documented as Vite only; other
unplugin adapters exist but are no-ops.
