+++ demo-lib/tsdown.config.ts
@@ -1,4 +1,4 @@
-import nostics from 'nostics/lib'
+import nosticsStrip from 'nostics/unplugin/strip-transform'
 import { defineConfig } from 'tsdown'
 
 export default defineConfig({
@@ -13,5 +13,5 @@ export default defineConfig({
     // can't enable due to https://github.com/microsoft/TypeScript/issues/58944#issuecomment-4213203205
     oxc: true,
   },
-  plugins: [nostics.rolldown()],
+  plugins: [nosticsStrip.rolldown()],
 })
