--- src/code-transform/dev.ts
@@ -1,2 +1,0 @@
-export { nosticsServer } from './server-plugin'
-export type { NosticsServerOptions } from './server-plugin'
