+++ src/unplugin/dev-server-collector.ts
@@ -1,0 +1,2 @@
+export { nosticsCollector } from './server-plugin'
+export type { NosticsCollectorOptions } from './server-plugin'
