+++ tsdown.config.ts
@@ -6,8 +6,8 @@ export default defineConfig({
     'index': 'src/index.ts',
     'formatters/ansi': 'src/formatters/ansi.ts',
     'formatters/json': 'src/formatters/json.ts',
-    'lib': 'src/code-transform/lib.ts',
-    'dev': 'src/code-transform/dev.ts',
+    'unplugin/strip-transform': 'src/unplugin/strip-transform.ts',
+    'unplugin/dev-server-collector': 'src/unplugin/dev-server-collector.ts',
     'reporters/dev': 'src/reporters/dev.ts',
     'reporters/node': 'src/reporters/node.ts',
     'reporters/fetch': 'src/reporters/fetch.ts',
